Reference #22 - Reports that the imported inactive orgs are being duplicate. (#25)
The issue is #comment when the pull:orgs command was ran when checking for duplicate orgs I was ignoring the ones that are already deactivated.

I also added an option on the pull:orgs command to remove duplicates. It looks first the 2nd instance of the first inactive org and deletes from that id on. to run it

`php artisan pull:orgs --org-cleanup`