Upgrade Laravel Framework & Livewire
-  Livewire v2.12.3 > v3.4.4

This is in preparation for a fresh install of Filament v3, which uses Livewire v3 under the hood.

Most upgrades were automatic via `livewire:upgrade` with a couple of exceptions:
- On HackgreenvilleTimeline.php:29, an old method "dispatchBrowserEvent()" was condensed into "dispatch()", and now takes args as named params instead of array keys.
- Livewire's config was published so the `app.blade.php` could stay where it was.

`@livewireStyles` and `@livewireScripts` also no longer need to be added manually.

- Laravel v10.14.1 > v10.34.2

This is due to Livewire v3 using Illuminate\Support\Stringable::isUrl() indirectly when rendering Livewire components. This method did not exist in earlier versions of Laravel 10.